Made S3 bucket name configurable via an envvar
This modification also eliminates access to all buckets across all environments; now the dev bucket should only be accessible by the dev proxy and so on, but this also allows us to maintain one URL that starts at resources/ instead of having one for each.  Hopefully this will prevent future confusion, it will just be a matter of using the correct S3 bucket for the intended environment.

A huge thank you to @jmcarp for the additional help and clearing up confusion!